requirements.txt
flask
bcrypt
uuid
python-dotenv
pymongo
pyjwt
flask_cors